requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
base58==2.1.1
billiard==4.2.1
bitarray==3.1.0
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cytoolz==1.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
ipfshttpclient==0.7.0
kombu==5.4.2
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
parsimonious==0.10.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2==2.9.10
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
pywin32==308
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20241016
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
varint==1.0.2
vine==5.1.0
wcwidth==0.2.13
web3==7.8.0
websockets==13.1
yarl==1.18.3
